+++ frontend/src/CatalogEntryForm.jsx
