+++ frontend/src/CatalogNavBar.jsx
