+++ frontend/src/DatasourceNav.jsx
@@ -1,4 +1,4 @@
-import CatalogNavBar from "./CatalogNavBar.js"
+import CatalogNavBar from "./CatalogNavBar.jsx"
 
 // Component: Tab-bar for datasources (grid, table, create etc)
 export default function DatasourceNav() {
