+++ frontend/src/ModelGrid.jsx
@@ -1,5 +1,5 @@
-import GridContents from "./GridContents.js"
-import ModelNav from "./ModelNav.js"
+import GridContents from "./GridContents.jsx"
+import ModelNav from "./ModelNav.jsx"
 
 import models from './catalog/data/models.json';
 import { model_thumbnails } from "./thumbnails.js"
