+++ frontend/src/ProjectTable.jsx
@@ -1,9 +1,9 @@
 import DataTable from 'react-data-table-component';
 
-import ProjectNav from "./ProjectNav.js"
+import ProjectNav from "./ProjectNav.jsx"
 import { project_thumbnails } from "./thumbnails.js"
-import { renderThumbnailForTable, TableCardDropdown } from "./table_helpers.js"
-import { TaskBadge } from "./badges.js";
+import { renderThumbnailForTable, TableCardDropdown } from "./table_helpers.jsx"
+import { TaskBadge } from "./badges.jsx";
 
 import projects from './catalog/data/projects.json';
