+++ frontend/src/table_helpers.jsx
